updated
added a tool tip for new button that toggles the first player to go in game.
adjusted some functions like change player and initcells for the new button functionality.
changed how the computer decides where it goes, depends on X and O now and if it is first player it choose two random locations first.